Extracted MenuBar to its own file
